requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.11.29
django-celery==3.3.1
django-excel==0.0.10
django-filter==2.0.0
djangorestframework==3.9.4
kombu==3.0.37
lml==0.0.9
Pillow==6.2.2
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.5.2
pytz==2020.1
texttable==1.6.2
